Move ReentrantLock to internal
